# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "MoneyDto": {
        "properties": {
          "amountMinor": { "example": "125000", "type": "string" },
          "currency": { "example": "IRR", "type": "string" }
        },
        "required": ["amountMinor", "currency"],
        "type": "object"
      },
...
      "PaymentResponseDto": {
        "properties": {
          "amount": { "$ref": "#/components/schemas/MoneyDto" },
          "createdAt": { "format": "date-time", "type": "string" },
          "id": { "format": "uuid", "type": "string" },
          "orderNumber": { "example": "HNY-2026-000123", "type": "string" },
          "paidAt": { "format": "date-time", "nullable": true, "type": "string" },
          "provider": { "type": "string" },
          "redirectUrl": { "nullable": true, "type": "string" },
          "status": {
            "enum": [
              "CREATED",
              "PENDING",
              "AUTHORIZED",
              "PAID",
              "FAILED",
              "CANCELLED",
              "EXPIRED",
              "REFUNDED",
              "PARTIALLY_REFUNDED"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "provider",
          "amount",
          "redirectUrl",
          "createdAt",
          "paidAt"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "code": { "example": "VALIDATION_FAILED", "type": "string" },
          "detail": { "example": "The request could not be processed.", "type": "string" },
          "errors": {
            "items": { "$ref": "#/components/schemas/ValidationIssueDto" },
            "type": "array"
          },
          "instance": { "example": "/v1/example", "type": "string" },
          "requestId": { "example": "018f5d36-7b89-7a67-bb7a-e8f9f5db7412", "type": "string" },
          "status": { "example": 422, "type": "number" },
          "title": { "example": "Request validation failed", "type": "string" },
          "type": {
            "example": "https://api.honey.invalid/problems/validation-failed",
            "type": "string"
          }
        },
        "required": ["type", "title", "status", "code", "instance", "requestId"],
        "type": "object"
      },
...
      "RefundResponseDto": {
        "properties": {
          "amount": { "$ref": "#/components/schemas/MoneyDto" },
          "completedAt": { "format": "date-time", "nullable": true, "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "id": { "format": "uuid", "type": "string" },
          "paymentId": { "format": "uuid", "type": "string" },
          "reason": { "type": "string" },
          "status": {
            "enum": ["REQUESTED", "PENDING", "COMPLETED", "FAILED", "CANCELLED"],
            "type": "string"
          }
        },
        "required": ["id", "paymentId", "amount", "reason", "status", "createdAt", "completedAt"],
        "type": "object"
      },
...
      "ValidationIssueDto": {
        "properties": {
          "code": { "example": "INVALID_VALUE", "type": "string" },
          "path": { "example": "field", "type": "string" }
        },
        "required": ["path", "code"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "description": "Opaque server-side session. Local HTTP development uses honey_session.",
        "in": "cookie",
        "name": "__Host-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Honey API contract. Cookie-authenticated unsafe requests require the X-CSRF-Token double-submit header.",
    "title": "Honey API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/payments/{paymentId}/refunds": {
      "post": {
        "description": "Rejects with a conflict unless the staff session has recently re-verified its TOTP via POST /v1/me/step-up. The server, not the caller, computes the refundable cap.",
        "operationId": "requestPaymentRefund",
        "parameters": [
          {
            "in": "path",
            "name": "paymentId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RefundResponseDto" } }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "Refund a paid payment (requires a recent step-up)",
        "tags": ["Payments"]
      }
    },
...
    "/v1/payments": {
      "post": {
        "description": "Creates or resumes a payment from the server-owned order total. The client may send only the order number and an Idempotency-Key; money and card fields are rejected.",
        "operationId": "startPayment",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Start (or resume) payment for an order",
        "tags": ["Payments"]
      }
    },
    "/v1/payments/{id}": {
      "get": {
        "description": "Returns the caller-owned payment projection. Another owner receives 404. Responses are private and must not be stored.",
        "operationId": "getPayment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Read one owner-scoped payment",
        "tags": ["Payments"]
      }
    },
    "/v1/payments/{id}/return": {
      "post": {
        "description": "The browser only signals \"check now\" — every field it carries in the query string is ignored; the server re-verifies its own stored provider reference (ADR-0022).",
        "operationId": "verifyPaymentReturn",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Verify the provider outcome after the customer returns from the gateway",
        "tags": ["Payments"]
      }
    },
    "/webhooks/payments/{provider}": {
      "post": {
        "description": "Accepts the exact raw body for signature verification. CSRF is not used; invalid signatures are persisted and ignored.",
        "operationId": "receivePaymentWebhook",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": { "example": "mock", "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Accepted; processing is idempotent and re-runnable." }
        },
        "summary": "Receive a signed provider webhook event",
        "tags": ["Payments"]
      }
    }
  },
  "servers": [],
...
}